Fix for u128 input/output rpc paramters in XYK pallet API
Actullay its a workaround for serde<->substrate issues.
Those seems to be fixed in v3.0.0. Trick is to use
NumberOrHex instead of u128 for input RPC api paramters.
NumberOrHex can be converted to u128 for XYK runtime api call
and is already shipped with proper serialization/deserialization
impl. What is more it allows for accepting request in both numerical
and hex/string form"