FIX(develop): Update workflow caching to fix faiiling github action used before (#778)
This pull request updates the caching mechanism to fix a failing GitHub
action that was used before. The changes include updating the
environment variables for the Google Cloud Platform (GCP)
authentication, downloading and uploading the cargo build cache, and
setting up the sccache action. These updates aim to improve the
performance and reliability of the GitHub action.

---------

Co-authored-by: goncer <gonza.cerro@gmail.com>